feat: InputSplitter 클래스로 로직 분리하기
StringCalculator 클래스에서 splitInput 메서드에 해당하는 로직이
꽤나 복잡해져서
InputSplitter 클래스로 분리하기

테스트 코드도 StringCalculatorTest 에서 InputSplitterTest 로 이동
